Moved player animator logic (#62)
