+++ Assets/Scripts/Entity/AbstractEntity.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using System;
-using UnityEngine;
-using UnityEngine.UI;
+﻿using UnityEngine;
 
 namespace Assets.Scripts.Entity
 {
@@ -24,28 +20,6 @@ namespace Assets.Scripts.Entity
             RB = GetComponent<Rigidbody2D>();
             Animator = GetComponent<Animator>();
             SR = GetComponent<SpriteRenderer>();
-        }
-
-        public void FixedUpdate()
-        {
-            Movement = Movement * Speed;
-            RB.velocity += Movement;
-
-            Animator.SetFloat("Speed", Math.Abs(RB.velocity.x));
-
-            Animator.SetBool("Falling", false);
-            Animator.SetBool("Jumping", false);
-
-            if (RB.velocity.y > 0.2f)
-                Animator.SetBool("Jumping", true);
-            else if (RB.velocity.y < -0.2f)
-                Animator.SetBool("Falling", true);
-
-            
-            if (RB.velocity.x > 0.1f)
-                SR.flipX = false;
-            else if (RB.velocity.x < -0.1f)
-                SR.flipX = true;
         }
     }
 }
