+++ Assets/Scripts/Entity/Enemy.cs
@@ -1,14 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Assets.Scripts.Entity
 {
     public class Enemy : AbstractEntity
     {
         public float HitPower = 100f;
-
-        new private void FixedUpdate(){ }
 
         public void OnCollisionEnter2D(Collision2D collidedWith) {
             if (collidedWith.gameObject.CompareTag("Player")) {
